Draggable lanes updated
