+++ app/src/lib/components/Board.svelte
@@ -4,7 +4,7 @@
 	import { Project } from '$lib/backend/projects';
 	import BranchDropzone from '$lib/branch/BranchDropzone.svelte';
 	import BranchLane from '$lib/branch/BranchLane.svelte';
-	import { cloneWithRotation } from '$lib/dragging/draggable';
+	import { cloneElement } from '$lib/dragging/draggable';
 	import { persisted } from '$lib/persisted/persisted';
 	import Icon from '$lib/shared/Icon.svelte';
 	import { getContext, getContextStore } from '$lib/utils/context';
@@ -94,12 +94,12 @@
 						e.stopPropagation();
 						return;
 					}
-					clone = cloneWithRotation(e.target);
+					clone = cloneElement(e.target as HTMLElement);
 					document.body.appendChild(clone);
 					// Get chromium to fire dragover & drop events
 					// https://stackoverflow.com/questions/6481094/html5-drag-and-drop-ondragover-not-firing-in-chrome/6483205#6483205
 					e.dataTransfer?.setData('text/html', 'd'); // cannot be empty string
-					e.dataTransfer?.setDragImage(clone, e.offsetX + 30, e.offsetY + 30); // Adds the padding
+					e.dataTransfer?.setDragImage(clone, e.offsetX, e.offsetY); // Adds the padding
 					dragged = e.currentTarget;
 					priorPosition = Array.from(dropZone.children).indexOf(dragged);
 				}}
