+++ app/src/lib/dragging/draggable.ts
@@ -1,82 +1,47 @@
 import { dropzoneRegistry } from './dropzone';
+import { getVSIFileIcon } from '$lib/ext-icons';
+import { pxToRem } from '$lib/utils/pxToRem';
+import { type CommitStatus } from '$lib/vbranches/types';
 import type { Draggable } from './draggables';
 
 export interface DraggableConfig {
 	readonly selector?: string;
 	readonly disabled?: boolean;
+	readonly label?: string;
+	readonly filePath?: string;
+	readonly sha?: string;
+	readonly dateAndAuthor?: string;
+	readonly commitType?: CommitStatus;
 	readonly data?: Draggable | Promise<Draggable>;
 	readonly viewportId?: string;
-	readonly extendWithClass?: string;
 }
 
-export function applyContainerStyle(element: HTMLElement) {
-	element.style.position = 'absolute';
-	element.style.top = '-9999px'; // Element has to be in the DOM so we move it out of sight
-	element.style.display = 'inline-block';
-	element.style.padding = '30px'; // To prevent clipping of rotated element
+function createElement(
+	tag: string,
+	classNames: string[],
+	textContent?: string,
+	src?: string
+): HTMLElement {
+	const el = document.createElement(tag);
+	el.classList.add(...classNames);
+	if (textContent) el.textContent = textContent;
+	if (src) (el as HTMLImageElement).src = src;
+	return el;
 }
 
-export function createContainerForMultiDrag(
-	children: Element[],
-	extendWithClass: string | undefined
-): HTMLDivElement {
-	const inner = document.createElement('div');
-	inner.style.display = 'flex';
-	inner.style.flexDirection = 'column';
-	inner.style.gap = '0.125rem';
-
-	children.forEach((child) => {
-		inner.appendChild(cloneWithPreservedDimensions(child, extendWithClass));
-	});
-	rotateElement(inner);
-
-	const container = document.createElement('div');
-	container.appendChild(inner);
-	applyContainerStyle(container);
-
-	return container;
-}
-
-export function cloneWithPreservedDimensions(node: any, extendWithClass: string | undefined) {
-	const clone = node.cloneNode(true) as HTMLElement;
-	clone.style.height = node.clientHeight + 'px';
-	clone.style.width = node.clientWidth + 'px';
-	clone.classList.remove('selected-draggable');
-
-	extendWithClass && clone.classList.add(extendWithClass);
-
-	return clone;
-}
-
-export function cloneWithRotation(node: any, extendWithClass: string | undefined = undefined) {
-	const container = document.createElement('div');
-	const clone = cloneWithPreservedDimensions(node, extendWithClass) as HTMLElement;
-	container.appendChild(clone);
-
-	// exclude all ignored elements from the clone
-	const ignoredElements = container.querySelectorAll('[data-remove-from-draggable]');
-	ignoredElements.forEach((element) => {
-		element.remove();
-	});
-
-	applyContainerStyle(container);
-
-	// Style the inner node so it retains the shape and then rotate
-	// TODO: This rotation puts a requirement on draggables to have
-	// an outer container, which feels extra. Consider refactoring.
-	rotateElement(clone);
-	return container as HTMLElement;
-}
-
-function rotateElement(element: HTMLElement) {
-	element.style.rotate = `${Math.floor(Math.random() * 3)}deg`;
-}
-
-export function draggable(node: HTMLElement, initialOpts: DraggableConfig) {
-	let opts = initialOpts;
+function setupDragHandlers(
+	node: HTMLElement,
+	opts: DraggableConfig,
+	createClone: (opts: DraggableConfig, selectedElements: HTMLElement[]) => HTMLElement,
+	params: {
+		handlerWidth: boolean;
+		maxHeight?: number;
+	} = {
+		handlerWidth: false
+	}
+) {
 	let dragHandle: HTMLElement | null;
-	let clone: HTMLElement | undefined;
-
+	let clone: HTMLElement;
 	let selectedElements: HTMLElement[] = [];
 
 	function handleMouseDown(e: MouseEvent) {
@@ -84,84 +49,71 @@ export function draggable(node: HTMLElement, initialOpts: DraggableConfig) {
 	}
 
 	function handleDragStart(e: DragEvent) {
-		let elt: HTMLElement | null = dragHandle;
+		e.stopPropagation();
 
-		while (elt) {
-			if (elt.dataset.noDrag !== undefined) {
-				e.stopPropagation();
-				e.preventDefault();
-				return false;
-			}
-			elt = elt.parentElement;
+		if (dragHandle && dragHandle.dataset.noDrag !== undefined) {
+			e.preventDefault();
+			return false;
 		}
 
-		// If the draggable specifies a selector then we check if we're dragging selected elements
 		if (opts.selector) {
-			// Checking for selected siblings in the parent of the parent container likely works
-			// for most use-cases but it was done here primarily for dragging multiple files.
-			const parentNode = node.parentNode?.parentNode;
-			selectedElements = parentNode
-				? Array.from(parentNode.querySelectorAll(opts.selector).values() as Iterable<HTMLElement>)
-				: [];
-
-			if (selectedElements.length > 0) {
-				clone = createContainerForMultiDrag(selectedElements, opts.extendWithClass);
-				// Dim the original element while dragging
-				selectedElements.forEach((element) => {
-					element.style.opacity = '0.5';
-				});
+			const parentNode = node.parentElement?.parentElement;
+			if (!parentNode) {
+				console.error('draggable parent node not found');
+				return;
 			}
+			selectedElements = Array.from(
+				parentNode.querySelectorAll(opts.selector) as NodeListOf<HTMLElement>
+			);
+			selectedElements = selectedElements.length ? selectedElements : [node];
 		}
 
-		if (!clone) {
-			clone = cloneWithRotation(node, opts.extendWithClass);
+		clone = createClone(opts, selectedElements);
+		if (params.handlerWidth) {
+			clone.style.width = node.clientWidth + 'px';
+		}
+		if (params.maxHeight) {
+			clone.style.maxHeight = pxToRem(params.maxHeight);
 		}
 
+		selectedElements.forEach((el) => (el.style.opacity = '0.5'));
 		document.body.appendChild(clone);
 
 		Array.from(dropzoneRegistry.values()).forEach((dropzone) => {
 			dropzone.register(opts.data);
 		});
 
-		// Get chromium to fire dragover & drop events
-		// https://stackoverflow.com/questions/6481094/html5-drag-and-drop-ondragover-not-firing-in-chrome/6483205#6483205
-		e.dataTransfer?.setData('text/html', 'placeholder copy'); // cannot be empty string
-		e.dataTransfer?.setDragImage(clone, e.offsetX + 30, e.offsetY + 30); // Adds the padding
-		e.stopPropagation();
+		if (e.dataTransfer) {
+			if (params.handlerWidth) {
+				e.dataTransfer.setDragImage(clone, e.offsetX, e.offsetY);
+			} else {
+				e.dataTransfer.setDragImage(clone, clone.offsetWidth - 20, 25);
+			}
+
+			// Get chromium to fire dragover & drop events
+			// https://stackoverflow.com/questions/6481094/html5-drag-and-drop-ondragover-not-firing-in-chrome/6483205#6483205
+			e.dataTransfer?.setData('text/html', 'd'); // cannot be empty string
+			e.dataTransfer.effectAllowed = 'uninitialized';
+		}
 	}
 
 	function handleDragEnd(e: DragEvent) {
-		if (clone) {
-			clone.remove();
-			clone = undefined;
-		}
-
-		// reset the opacity of the selected elements
-		selectedElements.forEach((element) => {
-			element.style.opacity = '1';
-		});
-
+		e.stopPropagation();
+		if (clone) clone.remove();
+		selectedElements.forEach((el) => (el.style.opacity = '1'));
 		Array.from(dropzoneRegistry.values()).forEach((dropzone) => {
 			dropzone.unregister();
 		});
-
-		e.stopPropagation();
 	}
 
-	const viewport = opts.viewportId ? document.getElementById(opts.viewportId) : null;
-	const triggerRange = 150;
-	const scrollSpeed = (viewport?.clientWidth || 500) / 2;
-	let lastDrag = new Date().getTime();
-
 	function handleDrag(e: DragEvent) {
+		e.preventDefault();
+		const viewport = opts.viewportId ? document.getElementById(opts.viewportId) : null;
 		if (!viewport) return;
-		if (new Date().getTime() - lastDrag < 500) return;
-		lastDrag = new Date().getTime();
-
+		const triggerRange = 150;
+		const scrollSpeed = (viewport.clientWidth || 500) / 2;
 		const viewportWidth = viewport.clientWidth;
 		const relativeX = e.clientX - viewport.getBoundingClientRect().left;
-
-		// Scroll horizontally if the draggable is near the edge of the viewport
 		if (relativeX < triggerRange) {
 			viewport.scrollBy(-scrollSpeed, 0);
 		} else if (relativeX > viewportWidth - triggerRange) {
@@ -189,12 +141,123 @@ export function draggable(node: HTMLElement, initialOpts: DraggableConfig) {
 	setup(opts);
 
 	return {
-		update(opts: DraggableConfig) {
+		update(newOpts: DraggableConfig) {
 			clean();
-			setup(opts);
+			setup(newOpts);
 		},
 		destroy() {
 			clean();
 		}
 	};
+}
+
+//////////////////////////
+//// COMMIT DRAGGABLE ////
+//////////////////////////
+
+export function createCommitElement(
+	commitType: CommitStatus | undefined,
+	label: string | undefined,
+	sha: string | undefined,
+	dateAndAuthor: string | undefined
+): HTMLDivElement {
+	const cardEl = createElement('div', ['draggable-commit']) as HTMLDivElement;
+	const labelEl = createElement('span', ['text-base-13', 'text-bold'], label);
+	const infoEl = createElement('div', ['draggable-commit-info', 'text-base-11']);
+	const shaEl = createElement('span', ['draggable-commit-info-text'], sha);
+	const dateAndAuthorEl = createElement('span', ['draggable-commit-info-text'], dateAndAuthor);
+
+	if (commitType) {
+		const indicatorClass = `draggable-commit-${commitType}`;
+		labelEl.classList.add('draggable-commit-indicator', indicatorClass);
+	}
+
+	cardEl.appendChild(labelEl);
+	infoEl.appendChild(shaEl);
+	infoEl.appendChild(dateAndAuthorEl);
+	cardEl.appendChild(infoEl);
+	return cardEl;
+}
+
+export function draggableCommit(node: HTMLElement, initialOpts: DraggableConfig) {
+	function createClone(opts: DraggableConfig) {
+		return createCommitElement(opts.commitType, opts.label, opts.sha, opts.dateAndAuthor);
+	}
+	return setupDragHandlers(node, initialOpts, createClone, {
+		handlerWidth: true
+	});
+}
+
+////////////////////////
+//// FILE DRAGGABLE ////
+////////////////////////
+
+export function createChipsElement(
+	childrenAmount: number,
+	label: string | undefined,
+	filePath: string | undefined
+): HTMLDivElement {
+	const containerEl = createElement('div', ['draggable-chip-container']) as HTMLDivElement;
+	const chipEl = createElement('div', ['draggable-chip']);
+	containerEl.appendChild(chipEl);
+
+	if (filePath) {
+		const iconEl = createElement(
+			'img',
+			['draggable-chip-icon'],
+			undefined,
+			getVSIFileIcon(filePath)
+		);
+		chipEl.appendChild(iconEl);
+	}
+
+	const labelEl = createElement('span', ['text-base-12'], label);
+	chipEl.appendChild(labelEl);
+
+	if (childrenAmount > 1) {
+		const amountTag = createElement(
+			'div',
+			['text-base-11', 'text-bold', 'draggable-chip-amount'],
+			childrenAmount.toString()
+		);
+		chipEl.appendChild(amountTag);
+	}
+
+	if (childrenAmount === 2) {
+		containerEl.classList.add('draggable-chip-two');
+	} else if (childrenAmount > 2) {
+		containerEl.classList.add('draggable-chip-multiple');
+	}
+
+	return containerEl;
+}
+
+export function draggableChips(node: HTMLElement, initialOpts: DraggableConfig) {
+	function createClone(opts: DraggableConfig, selectedElements: HTMLElement[]) {
+		return createChipsElement(selectedElements.length, opts.label, opts.filePath);
+	}
+	return setupDragHandlers(node, initialOpts, createClone);
+}
+
+////////////////////////
+//// HUNK DRAGGABLE ////
+////////////////////////
+
+export function cloneElement(node: HTMLElement) {
+	const cloneEl = node.cloneNode(true) as HTMLElement;
+
+	// exclude all ignored elements from the clone
+	const ignoredElements = Array.from(cloneEl.querySelectorAll('[data-remove-from-draggable]'));
+	ignoredElements.forEach((el) => el.remove());
+
+	return cloneEl;
+}
+
+export function draggableElement(node: HTMLElement, initialOpts: DraggableConfig) {
+	function createClone() {
+		return cloneElement(node);
+	}
+	return setupDragHandlers(node, initialOpts, createClone, {
+		handlerWidth: true
+	});
 }
